+++ examples/polymorphic/expected/UserOrPost/Node/Task.php
@@ -1,0 +1,16 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Spawnia\Sailor\Polymorphic\UserOrPost\Node;
+
+class Task extends \Spawnia\Sailor\TypedObject
+{
+    /** @var string */
+    public $id;
+
+    public function idTypeMapper(): callable
+    {
+        return new \Spawnia\Sailor\Mapper\DirectMapper();
+    }
+}